requirements.txt
streamlit
scikit-learn
pandas
joblib
matplotlib
seaborn
pytesseract
Pillow
